requirements.txt
python-dotenv
selenium
webdriver-manager

